Third implementation.
For each node, instead of publishing the same message to the channel,
the messages are separated into two streams:

1. periodically published msg to 'all' by a timer;
2. immedietaly responsed msg to 'someone' on receiving a msg sent to 'all'.

On receiving a message sent to 'all', the node info in the message will
be added to db. Followed by a response with a designated receiver and the
sender's node info.

The node sent the message to 'all' then gets the response and add the
node info in the response to its db.

These two kinds of messages used to be one, now got separated.

And the first message stream will stop once the node found itself is in
a cluster. Then it will only send response messages.