Remove interface for start publishing.
pub is started with erlang_node_discovery_pubsub, after sub.
It is for two cases:

1. When there is only one node in the cluster, a timer for publising
is periodically triggered. If unfortunately pubsub got restarted for
some reason, there is no way to restart publish timer again.

2. When two nodes already connected two each other, meaning in a
cluster, both of them will stop the timer for publishing. At this time
when one of them (call it A) failed to receive channel message for some
network issue and a third node (call it C) joins the cluster. Node A
won't know about it. If it comes back again normally, it will got
published messages from the subscribed channel (there is flaw here).
But if it got restarted during the network issue time, it will never
know about node C. And Node C will never know about node A, unless send
a message to the channel at initilization phase.

When the two cases combine togather, the solution is start publishing
right after subscribing in the pubsub module. Remove the pub interface.